catalogo_facil/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catalogo_facil"
version = "0.1.0"
description = "Sistema em Flask para catálogo de produtos e registro de vendas"
requires-python = ">=3.11"
dependencies = [
    "Flask==3.1.0",
    "Flask-SQLAlchemy==3.1.1",
    "Pillow==12.0.0",
    "matplotlib==3.10.8",
    "reportlab==4.4.4",
    "waitress==3.0.2",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
app = [
    "templates/**/*",
    "static/**/*",
]